% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qic.R
\name{qic}
\alias{qic}
\title{Quasi-likelihood under the Independence Model Criterion}
\usage{
qic(object, type = "pan", tol = .Machine$double.eps, envir = parent.frame())
}
\arguments{
\item{object}{An object of class \code{"gee"} representing the fit.}

\item{type}{A character string specifying the type of penalty to calculate.
The following are permitted: \code{"pan"}, \code{"hh"}, \code{"u"}.
Default is \code{"pan"}.}

\item{tol}{A numeric value specifying the tolerance when calculating the
inverse of the model-based covariance matrix.
Default is \code{.Machine$double.eps}.}

\item{envir}{An environment object to which find the \code{"gee"} object.}
}
\value{
A named numeric vector containing the QIC value for the \code{"gee"}
object under consideration. The name of the value is based on \code{type},
where \code{"QIC"} denotes \code{type = "pan"}, \code{"QIChh"} denotes
\code{type = "hh"}, and \code{"QICu"} denotes \code{type = "u"}.
}
\description{
Calculating the quasi-likelihood under the independence model
criterion (QIC) based on Pan (2001a).
}
\details{
The calculation of QICu proposed by Pan(2001) is calculated when it
is assumed that they working correlation structure is properly specified.
This value can be obtained by specifying \code{type = "u"}. Specifying
\code{type = "hh"} calculates the penalty based on Hardin and Hilbe (2012).
More information on this penalty can be found at \code{?gee::cic}.
}
\examples{
# Marginal analysis of random effects model for wool
data(warpbreaks)
fit <- gee::gee(
 breaks ~ tension, id = wool, data = warpbreaks, corstr = "AR-M"
)
gee::qic(fit, type = "pan")
gee::qic(fit, type = "hh")
gee::qic(fit, type = "u")

}
\references{
Pan, W. (2001a). \emph{Akaike's information criterion in generalized
estimating equations}. Biometrics, 57, 120-125.\cr
Hardin, J.W.  and Hilbe, J.M. (2012). \emph{Generalized
Estimating Equations, 2nd Edition}, Chapman and Hall/CRC: New York. \cr
Hin, L.-Y. and Wang, Y-G. (2009). \emph{Working-correlation-structure
identification in generalized estimating equations},
Statistics in Medicine 28: 642-658.
}
\seealso{
\code{\link{gee::cic}}, \code{\link{gee::quasilik}}
}
