% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qic.R
\name{quasilik}
\alias{quasilik}
\title{Quasi-likelihood under the Indepence Model}
\usage{
quasilik(object)
}
\arguments{
\item{object}{An object of class \code{"gee"} representing the fit.}
}
\value{
A numeric vector containing the quasi-likelihood under the
independence model value for the \code{"gee"} object under consideration.
}
\description{
Calculating the quasi-likelihood under the independence model
based on estimating equations proposed by Liang and Zeger (1986).
}
\examples{
# Marginal analysis of random effects model for wool
data(warpbreaks)
fit <- gee::gee(
 breaks ~ tension, id = wool, data = warpbreaks, corstr = "AR-M"
)
gee::quasilik(object = fit)

}
\references{
Liang, K.Y. and Zeger, S.L. (1986) Longitudinal data analysis using
generalized linear models. \emph{Biometrika}, 73 13-22.
}
